+++ cli-tool/app/commands/export.rb
@@ -28,6 +28,7 @@ module Taylor
             --help\t\t\tDisplays this message
             --dry-run\t\t\tJust display the export command and don't run it
             --export_directory directory\tWhat directory do you want your exports (defaults to ./exports)
+            --build_cache directory\tWhere do you want to store build cache (defaults to nil)
         STR
       end
 
@@ -43,11 +44,12 @@ module Taylor
         parser = OptParser.new do |opts|
           opts.on(:help,             :bool,   false)
           opts.on(:dry_run,          :bool,   false)
+          opts.on(:export_directory, :string, options.fetch(:export_directory, './exports'))
+          opts.on(:build_cache,      :string)
         end
         parser.parse(argv, true)
 
         @options = parser.opts
-        @options[:export_directory] = options.fetch(:export_directory, './exports')
       end
 
       def check_in_taylor_project!
@@ -65,9 +67,23 @@ module Taylor
         end
       end
 
+      def create_build_cache_folder
+        return if @options[:dry_run]
+        return if @options[:build_cache].nil?
+
+        unless File.exists?(options[:build_cache]) &&
+            File.directory?(options[:build_cache])
+          Dir.mkdir(options[:build_cache])
+        end
+      end
+
       def docker_build
         command = 'docker run -u $(id -u ${USER}):$(id -g ${USER})'
-        command << " --mount type=bind,source=#{Dir.pwd},target=/app/game/"
+        command << " --mount type=bind,source=#{Dir.pwd},target=/app/game"
+        command << " --mount type=bind,source=#{File.expand_path(@options[:export_directory])},target=/app/game/exports"
+        unless @options[:build_cache].nil?
+          command << " --mount type=bind,source=#{@options[:build_cache]},target=/app/taylor/build/"
+        end
         command << " hellrok/taylor:v#{TAYLOR_VERSION}"
 
         if options[:dry_run]